pali: Restructuring changes
Changes include:
    - restructuring code structure
    - Fix LICENSE
    - Move version in __init__.py
