+++ pali/__init__.py
@@ -1,0 +1,2 @@
+name = "pali"
+__version__ = '0.0.2'
