+++ pali/common.py
