+++ pali/constants.py
