+++ pali/pipeline.py
