+++ setup.py
@@ -5,15 +5,20 @@ import setuptools
 with open("README.md", "r") as fh:
     long_description = fh.read()
 
+version = {}
+with open("pali/__init__.py") as fp:
+    exec(fp.read(), version)
+
+
 setuptools.setup(
     name="pali",
-    version="0.0.1",
+    version=version['__version__'],
     author="Vipin Sharma",
     author_email="sh.vipin@gmail.com",
     description="A simple ThreadPool library",
     long_description=long_description,
     long_description_content_type="text/markdown",
-    url="https://github.com/gitvipin/hemis/tree/master/projects/pali",
+    url="https://github.com/gitvipin/pali",
     packages=setuptools.find_packages(),
     classifiers=[
         "Programming Language :: Python :: 2",
