+++ tests/test_worker_pool.py
@@ -5,8 +5,8 @@ Worker Pool testcases.
 
 import unittest
 
-import src.task as task
-import src.worker as worker
+import pali.task as task
+import pali.worker as worker
 
 
 class MyTask(task.Task):
